app/services/claim_repair_service.py
# app/services/claim_repair_service.py
"""
Placeholder for AI-powered claim repair suggestions.
This service would interact with an external AI model or a more complex internal logic
to suggest fixes for common claim errors.
"""
from app.utils.logging_config import get_logger
# from app.utils.error_handler import ExternalServiceError # If calling an external API

logger = get_logger('app.services.claim_repair')

class ClaimRepairService:
    """
    Provides suggestions for repairing failed claims.
    """
    def __init__(self, config: dict):
        self.config = config.get('services', {}).get('claim_repair', {})
        self.enabled = self.config.get('enabled', False)
        # self.api_url = self.config.get('repair_suggestion_api_url') # If using external API

        if self.enabled:
            logger.info("Claim Repair Service is enabled.")
            # Initialize any clients or models here if needed
            # if self.api_url:
            #     logger.info(f"Claim repair suggestions will be fetched from: {self.api_url}")
        else:
            logger.info("Claim Repair Service is disabled.")


    def get_repair_suggestions(self, claim_id: str, failed_claim_details: dict) -> List[dict]:
        """
        Generates or fetches repair suggestions for a given failed claim.

        Args:
            claim_id (str): The ID of the failed claim.
            failed_claim_details (dict): A dictionary containing details about the failure
                                         (e.g., error codes, messages, relevant claim data).

        Returns:
            List[dict]: A list of suggestions, where each suggestion is a dict
                        (e.g., {"field_to_correct": "facility_id", 
                                "suggested_value": "FAC002", 
                                "confidence": 0.85,
                                "reasoning": "Provided facility ID was not found. Common typo for 'FAC001'."})
        """
        if not self.enabled:
            logger.debug(f"Claim repair service disabled. No suggestions for claim {claim_id}.")
            return []

        logger.info(f"Attempting to get repair suggestions for claim {claim_id}.")
        suggestions = []

        # --- Placeholder Logic ---
        # This would be replaced with actual AI model calls or complex rule-based heuristics.
        
        error_codes = failed_claim_details.get('ErrorCodes', "")
        error_messages = failed_claim_details.get('ErrorMessages', "")

        if "VAL_FAC_001" in error_codes: # Example: Facility ID not found
            suggestions.append({
                "field_to_correct": "facility_id",
                "current_value": failed_claim_details.get('FacilityId', 'N/A'),
                "suggested_action": "Verify facility ID against master list. Check for typos.",
                "suggested_value_options": ["FAC001_VALID", "FAC002_VALID"], # Example valid options
                "confidence": 0.70,
                "reasoning": "Error code VAL_FAC_001 indicates facility ID was not found or inactive."
            })

        if "Patient Account Number is missing" in error_messages:
            suggestions.append({
                "field_to_correct": "patient_account_number",
                "current_value": failed_claim_details.get('PatientAccountNumber', ''),
                "suggested_action": "Ensure patient account number is provided and correct.",
                "confidence": 0.90,
                "reasoning": "Patient account number is a required field for billing."
            })
        
        # Example: If using an external API
        # if self.api_url:
        #     try:
        #         # import httpx
        #         # response = httpx.post(self.api_url, json={"claim_id": claim_id, "errors": failed_claim_details})
        #         # response.raise_for_status()
        #         # api_suggestions = response.json().get("suggestions", [])
        #         # suggestions.extend(api_suggestions)
        #         pass # Placeholder for API call
        #     except Exception as e:
        #         logger.error(f"Error calling external claim repair API for claim {claim_id}: {e}", exc_info=True)
        #         # raise ExternalServiceError(f"Failed to get repair suggestions from API: {e}", service_name="ClaimRepairAI")

        if not suggestions:
            logger.info(f"No specific repair suggestions generated for claim {claim_id} based on current logic.")
        else:
            logger.info(f"Generated {len(suggestions)} repair suggestions for claim {claim_id}.")
            
        return suggestions

if __name__ == '__main__':
    from app.utils.logging_config import setup_logging
    import yaml # For dummy config

    setup_logging()
    logger.info("--- Claim Repair Service Test ---")
    
    dummy_config = {
        'services': {
            'claim_repair': {
                'enabled': True,
                # 'repair_suggestion_api_url': 'http://mockrepair.api/suggest'
            }
        }
    }
    # If you want to test with config.yaml, ensure it has the services.claim_repair section.
    # For isolated test:
    repair_service = ClaimRepairService(config=dummy_config)

    test_failed_claim_1 = {
        "OriginalClaimId": "FAIL001",
        "FacilityId": "UNKNOWN_FAC",
        "ErrorCodes": "VAL_FAC_001, VAL_DATE_001",
        "ErrorMessages": "Facility ID 'UNKNOWN_FAC' not found.\nClaim service date is missing."
    }
    suggestions1 = repair_service.get_repair_suggestions("FAIL001", test_failed_claim_1)
    logger.info(f"Suggestions for FAIL001: {suggestions1}")

    test_failed_claim_2 = {
        "OriginalClaimId": "FAIL002",
        "PatientAccountNumber": "",
        "ErrorCodes": "VAL_PAT_001",
        "ErrorMessages": "Patient Account Number is missing."
    }
    suggestions2 = repair_service.get_repair_suggestions("FAIL002", test_failed_claim_2)
    logger.info(f"Suggestions for FAIL002: {suggestions2}")
    
    logger.info("Claim Repair Service test finished.")